Refactor EntryLogger class
- move EntryLogManagerBase from EntryLogManager to its own class.
